requirements.txt
flask~=3.1.0
psycopg2-binary~=2.9.10
Flask-Migrate~=4.1.0
alembic~=1.14.1